backend/requirements.txt
flask
flask-cors
python-dotenv
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langchain-groq==0.1.9
unstructured==0.15.0
unstructured[pdf]==0.15.0
faiss-cpu
transformers==4.43.3
sentence-transformers==3.0.1
nltk==3.8.1
